requirements.txt
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.4.2
Django==2.0.2
django-crispy-forms==1.7.0
django-filter==1.1.0
django-heroku==0.2.0
django-registration-redux==2.2
djangorestframework==3.7.7
flake8==3.5.0
gunicorn==19.7.1
idna==2.6
Markdown==2.6.11
mccabe==0.6.1
pew==1.1.2
pipenv==9.0.3
psycopg2==2.7.4
pycodestyle==2.3.1
pyflakes==1.6.0
pytz==2018.3
requests==2.18.4
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
waitress==1.1.0
whitenoise==3.3.1
psycopg2>=2.7,<2.8 --no-binary psycopg2
